refactor(ReceiverLink): move disposition handling to ReceiverLink
