+++ lib/link.js
@@ -119,22 +119,6 @@ Link.prototype.detach = function() {
   return detachPromise;
 };
 
-Link.prototype.accept = function(message) {
-  var range = u.dispositionRange(message);
-  this._sendDisposition(_.defaults(range, {
-    settled: true,
-    state: new DeliveryState.Accepted()
-  }));
-};
-
-Link.prototype.reject = function(message, reason) {
-  var range = u.dispositionRange(message);
-  this._sendDisposition(_.defaults(range, {
-    settled: true,
-    state: new DeliveryState.Rejected({ error: reason })
-  }));
-};
-
 // private api
 Link.prototype._attachReceived = function(attachFrame) {
   this.linkSM.attachReceived();
@@ -211,18 +195,6 @@ Link.prototype._detached = function(frame) {
 
   // @todo: Session should listen for Link.Detached
   this.session._linkDetached(this);
-};
-
-Link.prototype._sendDisposition = function(options) {
-  var dispositionOptions = _.defaults(options, {
-    role: constants.linkRole.receiver,
-    channel: this.session.channel,
-    handle: this.handle
-  });
-
-  this.session.connection.sendFrame(
-    new DispositionFrame(dispositionOptions)
-  );
 };
 
 module.exports = Link;
