+++ lib/receiver_link.js
@@ -6,9 +6,12 @@ var _ = require('lodash'),
 
     errors = require('./errors'),
     constants = require('./constants'),
+    u = require('./utilities'),
 
     FlowFrame = require('./frames/flow_frame'),
+    DispositionFrame = require('./frames/disposition_frame'),
 
+    DeliveryState = require('./types/delivery_state'),
     Link = require('./link');
 
 function ReceiverLink(session, handle, linkPolicy) {
@@ -52,6 +55,22 @@ ReceiverLink.prototype.addCredits = function(credits, flowOptions) {
   });
 };
 
+ReceiverLink.prototype.accept = function(message) {
+  var range = u.dispositionRange(message);
+  this._sendDisposition(_.defaults(range, {
+    settled: true,
+    state: new DeliveryState.Accepted()
+  }));
+};
+
+ReceiverLink.prototype.reject = function(message, reason) {
+  var range = u.dispositionRange(message);
+  this._sendDisposition(_.defaults(range, {
+    settled: true,
+    state: new DeliveryState.Rejected({ error: reason })
+  }));
+};
+
 // private API
 ReceiverLink.prototype._flowReceived = function(flowFrame) {
   this.drain = flowFrame.drain;
@@ -59,6 +78,16 @@ ReceiverLink.prototype._flowReceived = function(flowFrame) {
   this.emit(Link.CreditChange, this);
 };
 
+ReceiverLink.prototype._sendDisposition = function(options) {
+  var dispositionOptions = _.defaults(options, {
+    role: constants.linkRole.receiver,
+    channel: this.session.channel,
+    handle: this.handle
+  });
 
+  this.session.connection.sendFrame(
+    new DispositionFrame(dispositionOptions)
+  );
+};
 
 module.exports = ReceiverLink;
